Refactoring: class StepaArray separated from TTrack, because it doesn't use TTrack inner features
